A Data Structure Based project (group of 4). The Aim is to facilitate a tourist to search and reach his destination. This project includes t
PROBLEM  STATEMENT:
A  “city help center”   is a resource center for anyone new to the city
of Delhi. It is a full and complete guide for gathering first hand
information about all that’s the city has to offer.
The design consists of classes of various categories of places landmarks
and some basic information about them. Further other sub classes to them
provide the insight into each category with more detailed specifications
of each destination.
The overall model has two different accessors to for the application.
One of the Accessor can alter the database and thereby add or remove
destinations as per the developments in the city. The functions of this
accessor are password protected. The other Accessor is the main user for
whom this model was made. He can view the guide of the city, search for
information of a desired destination. Thus the sub classes of the
category class has the required functions according to the accessor for
implementing the process of adding a destination, removing the
destination, searching a destination or locating a destination.
All the classes also has functions to allow the user to go back to the
previous step at any point of time thus making it a user friendly
application.
This whole Scenario is implemented using suitable concepts of object
oriented programming in c++.